@@ -580,11 +580,12 @@ class _MessageWidgetState extends State<MessageWidget>
 
   bool get isOnlyEmoji => widget.message.text?.isOnlyEmoji == true;
 
-  bool get hasNonUrlAttachments =>
-      widget.message.attachments.where((it) => it.titleLink == null).isNotEmpty;
+  bool get hasNonUrlAttachments => widget.message.attachments
+      .where((it) => it.titleLink == null || it.type == 'giphy')
+      .isNotEmpty;
 
-  bool get hasUrlAttachments =>
-      widget.message.attachments.any((it) => it.titleLink != null) == true;
+  bool get hasUrlAttachments => widget.message.attachments
+      .any((it) => it.titleLink != null && it.type != 'giphy');
 
   bool get showBottomRow =>
       showThreadReplyIndicator ||
@@ -1156,7 +1157,9 @@ class _MessageWidgetState extends State<MessageWidget>
     final attachmentGroups = <String, List<Attachment>>{};
 
     widget.message.attachments
-        .where((element) => element.titleLink == null && element.type != null)
+        .where((element) =>
+            (element.titleLink == null && element.type != null) ||
+            element.type == 'giphy')
         .forEach((e) {
       if (attachmentGroups[e.type] == null) {
         attachmentGroups[e.type!] = [];
